+++ Changelog.md
@@ -13,6 +13,8 @@ Showroom View: Component presentation view for showcasing the library on a devic
 - Logos and iso added to the brand components
 - Added TopBars
 - Squared and Rectangle button component added.
+- added Toast Component
+
 ## [0.0.1] - 2025-10-7
 ### Added
 Initial Setup: Creation of the Kotlin Multiplatform library module (toteatds) with support for Android, iOS, and JVM.